tiki/requirements.txt
pandas
requests
confluent-kafka
psycopg2-binary


